backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pymongo==4.3.3
openai==0.27.8
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0
pyttsx3==2.90
aiofiles==23.1.0
